Move find under pkg/
